requirements.txt
amqp==5.1.1
apted==1.0.3
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
bleach==4.1.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
commonmark==0.9.1
cssselect==1.1.0
dj-database-url==0.5.0
Django==3.2.12
django-cors-headers==3.11.0
djangorestframework==3.13.1
docutils==0.18.1
gunicorn==20.1.0
html-matcher==0.0.8
html-similarity==0.3.3
idna==3.3
importlib-metadata==4.11.3
keyring==23.5.0
kombu==5.2.4
lxml==4.8.0
numpy==1.22.3
packaging==21.3
pandas==1.4.2
parsel==1.6.0
pkginfo==1.8.2
prompt-toolkit==3.0.29
psycopg2==2.9.3
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.3
readme-renderer==34.0
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.2.0
six==1.16.0
spmf==1.4
sqlparse==0.4.2
twine==4.0.0
typing_extensions==4.1.1
urllib3==1.26.7
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
XlsxWriter==3.0.3
zipp==3.8.0
